Yemen Governorate (adm1) and District (adm2) Administrative Boundaries

A list of length 2 containing geographic administrative boundary data for Yemen, stored as simple features (sf) objects.

Usage

sf_yem

Format

named list of sf objects

Details

Each admin level can be joined to the example df_ll dataset with a join by specification of c("pcode" = "adm1_pcode") and c("pcode" = "adm2_pcode") respectively. These should be passed as the join_by field in each geo_data specification passed to place_ui and place_server.

Examples

sf_yem$adm1
#> Simple feature collection with 22 features and 8 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 41.81479 ymin: 12.10665 xmax: 54.5382 ymax: 19
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    adm0_iso3 adm0_sub     adm0_name                   adm1_name pcode      lon
#> 1        YEM      ALL Yemen / اليمن                    Ibb / اب  YE11 44.18040
#> 2        YEM      ALL Yemen / اليمن                Abyan / ابين  YE12 46.23681
#> 3        YEM      ALL Yemen / اليمن Sana'a City / امانة العاصمه  YE13 44.17556
#> 4        YEM      ALL Yemen / اليمن          Al Bayda / البيضاء  YE14 45.27953
#> 5        YEM      ALL Yemen / اليمن                 Ta'iz / تعز  YE15 43.78494
#> 6        YEM      ALL Yemen / اليمن             Al Jawf / الجوف  YE16 45.59372
#> 7        YEM      ALL Yemen / اليمن                Hajjah / حجه  YE17 43.20314
#> 8        YEM      ALL Yemen / اليمن       Al Hodeidah / الحديده  YE18 43.21762
#> 9        YEM      ALL Yemen / اليمن          Hadramawt / حضرموت  YE19 48.63126
#> 10       YEM      ALL Yemen / اليمن               Dhamar / ذمار  YE20 44.30275
#>         lat adm1_pop                       geometry
#> 1  14.08362  3117999 MULTIPOLYGON (((44.08076 14...
#> 2  13.63099   619003 MULTIPOLYGON (((46.29563 14...
#> 3  15.40928  3981000 MULTIPOLYGON (((44.3338 15....
#> 4  14.32562   830001 MULTIPOLYGON (((44.72676 14...
#> 5  13.29334  3487612 MULTIPOLYGON (((43.41111 12...
#> 6  16.59384   645000 MULTIPOLYGON (((46.34001 17...
#> 7  16.09100  2415001 MULTIPOLYGON (((42.80233 15...
#> 8  14.78216  3653999 MULTIPOLYGON (((42.6918 13....
#> 9  16.52279  1618329 MULTIPOLYGON (((50.83766 16...
#> 10 14.54205  2170000 MULTIPOLYGON (((44.70527 14...
sf_yem$adm2
#> Simple feature collection with 335 features and 9 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 41.81479 ymin: 12.10665 xmax: 54.5382 ymax: 19
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    adm0_iso3 adm0_sub     adm0_name adm1_name                  adm2_name  pcode
#> 1        YEM      ALL Yemen / اليمن  Ibb / اب            Al Qafr / القفر YE1101
#> 2        YEM      ALL Yemen / اليمن  Ibb / اب               Yarim / يريم YE1102
#> 3        YEM      ALL Yemen / اليمن  Ibb / اب         Ar Radmah / الرضمه YE1103
#> 4        YEM      ALL Yemen / اليمن  Ibb / اب       An Nadirah / النادره YE1104
#> 5        YEM      ALL Yemen / اليمن  Ibb / اب         Ash Sha'ir / الشعر YE1105
#> 6        YEM      ALL Yemen / اليمن  Ibb / اب          As Saddah / السده YE1106
#> 7        YEM      ALL Yemen / اليمن  Ibb / اب      Al Makhadir / المخادر YE1107
#> 8        YEM      ALL Yemen / اليمن  Ibb / اب             Hobeish / حبيش YE1108
#> 9        YEM      ALL Yemen / اليمن  Ibb / اب Hazm Al Odayn / حزم العدين YE1109
#> 10       YEM      ALL Yemen / اليمن  Ibb / اب Far' Al Odayn / فرع العدين YE1110
#>         lon      lat adm2_pop                       geometry
#> 1  44.04029 14.32486   152248 MULTIPOLYGON (((43.82405 14...
#> 2  44.37298 14.28908   257878 MULTIPOLYGON (((44.2682 14....
#> 3  44.55286 14.21038   111778 MULTIPOLYGON (((44.4855 14....
#> 4  44.48370 14.06157   108243 MULTIPOLYGON (((44.4946 14....
#> 5  44.36793 14.03340    58233 MULTIPOLYGON (((44.307 14.0...
#> 6  44.41946 14.16679   119880 MULTIPOLYGON (((44.44934 14...
#> 7  44.20557 14.14071   166083 MULTIPOLYGON (((44.22156 14...
#> 8  44.10425 14.11770   152918 MULTIPOLYGON (((44.01363 14...
#> 9  43.83321 14.11171   115630 MULTIPOLYGON (((44.05006 14...
#> 10 43.78992 13.95782   130477 MULTIPOLYGON (((43.80641 14...