Time module

Visualise data over time with an interactive 'epicurve'.

Usage

time_ui(
  id,
  date_vars,
  group_vars = NULL,
  title = "Time",
  icon = bsicons::bs_icon("bar-chart-line-fill"),
  opts_btn_lab = "options",
  date_lab = "Date axis",
  date_int_lab = "Date interval",
  day_week_month_labs = c("Day", "Week", "Month"),
  groups_lab = "Group data by",
  bar_stacking_lab = "Bar stacking",
  cumul_data_lab = "Show cumulative data?",
  n_lab = "N patients",
  ratio_line_lab = NULL,
  full_screen = TRUE
)

time_server(
  id,
  df_ll,
  date_vars,
  group_vars = NULL,
  y_lab = "Patients",
  n_lab = "N patients",
  ratio_var = NULL,
  ratio_lab = NULL,
  ratio_numer = NULL,
  ratio_denom = NULL,
  filter_info = shiny::reactiveVal()
)

Arguments

id

Module id. Must be the same in both the UI and server function to link the two.

date_vars

named character vector of date variables for the date axis input. Names are used as variable labels.

group_vars

Character vector of categorical variable names. If provided, a select input will appear in the options dropdown allowing for data groups to be visualised as stacked bars on the epicurve. If named, names are used as variable labels.

title

header title for the card.

icon

The icon to display next to the title.

opts_btn_lab

text label for the dropdown menu button.

date_lab

text label for the date variable input.

date_int_lab

text label for the date interval input.

day_week_month_labs

character vector with text labels for day, week and month, respectively.

groups_lab

text label for the grouping variable input.

bar_stacking_lab

text label for bar stacking option.

cumul_data_lab

text label for cumulative data option.

n_lab

The label for the raw count variable.

ratio_line_lab

text label for the ratio line input. If not supplied the input is not included.

full_screen

Add button to card to with the option to enter full screen mode?

df_ll

Data frame or tibble of patient level linelist data. Can be either a shiny reactive or static dataset.

y_lab

The label for y-axis of chart.

ratio_var

Character string of variable name to use for ratio calculation.

ratio_lab

The label to describe the computed ratio i.e. 'CFR' for case fatality ratio.

ratio_numer

Value(s) in ratio_var to be used for the ratio numerator i.e. 'Death'.

ratio_denom

Values in ratio_var to be used for the ratio denominator i.e. c('Death', 'Recovery').

filter_info

If contained within an app using filter_server(), supply the filter_info element returned by that function here as a shiny reactive to add filter information to chart exports.

Value

the module server function returns any point click event data of the highchart. see highcharter::hc_add_event_point for details.